% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celltype}
\alias{celltype}
\title{Cell types as labels of example scRNA-seq dataset(exprdata)}
\format{A vector contains 4 cell types with its length 950}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE89232}
}
\usage{
celltype
}
\description{
A vector contains the cell types  as labels for cells in
example scRNA-seq dataset(exprdata)
}
\keyword{datasets}
