% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celltype_toy}
\alias{celltype_toy}
\title{toy cell type vector with 3 cell types generated for 5 cells in toy dataset}
\usage{
data(celltype_toy)
}
\description{
The cell type vector is generated from the following code
}
\details{
celltype = as.factor(sample(c(1,2,3),5,replace = TRUE))
}
\keyword{data}
