% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disperPlot}
\alias{disperPlot}
\title{Reference dataset(disperPlot)}
\format{A 53198 x 4 data frame}
\usage{
disperPlot
}
\description{
A dataframe contains all the gene-wise dispersion estimates loess curve for 14 datasets we benchmarked in Figure 2.a
}
\details{
The variable in the columns are:
fitted_dispersion: the log value of gene-wise dispersion after fitting a loess curve with respect to TPM value.
Note that the genes at the top 2.5% and buttom 2.5% of the TPM value has been removed
meantpm is average tpm value calculated per gene
dataset are nams for datasets
variance is  gene selection method, here is HEG vs HVG
}
\keyword{datasets}
