% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprdata}
\alias{exprdata}
\title{Processed scRNA-seq expression profile of dendritic dataset}
\format{A matrix with interger entries and have 2000 rows and 1059 columns}
\source{
\url{http://imlspenticton.uzh.ch/robinson_lab/conquer/data-mae/GSE48968-GPL13112.rds}
}
\usage{
exprdata
}
\description{
A gene count matrix with its rows are genes and columns are cells. The entry
of the matrix stands for the abundance of mRNA  in certain cell.
}
\keyword{datasets}
