% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFA.R
\name{getGeneExpr}
\alias{getGeneExpr}
\title{Function to extract gene expression matrix from input observation matrix}
\usage{
getGeneExpr(scData)
}
\arguments{
\item{scData}{can be a raw count matrix,
in which rows are genes and columns are cells;
can be a seurat object; can be a SingleCellExperiment object}
}
\value{
a raw expression matrix in which rows are genes and columns are cells.
}
\description{
Function to extract gene expression matrix from input observation matrix
}
\examples{
scData = matrix(rpois(15,1),3,5)

GeneExpr = getGeneExpr(scData)
}
\keyword{export}
