% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFA.R
\name{getLoading}
\alias{getLoading}
\title{Function to get low dimensional loading matrix}
\usage{
getLoading(modelEnv)
}
\arguments{
\item{modelEnv}{output environment variable}
}
\value{
\eqn{A}: \eqn{G} by \eqn{K} compressed feature space
}
\description{
Function to get low dimensional loading matrix
}
\examples{
GeneExpr = matrix(rpois(15,1),3,5)
bfa_model = scbfa(scData = GeneExpr,X = NULL,numFactors =2)
A = getLoading(bfa_model)
}
\keyword{export}
