% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFA.R
\name{gradient_chunk}
\alias{gradient_chunk}
\title{Calculate gradient of the negative log likelihood,
used for calls to the optim() function.}
\usage{
gradient_chunk(parameters, modelEnv)
}
\arguments{
\item{parameters}{Vectorized parameter space.}

\item{modelEnv}{Environment variable contains parameter space,
and global variables such as N,G,C,detection matrix B, etc}
}
\value{
Vectorized gradient
}
\description{
Calculate gradient of the negative log likelihood,
used for calls to the optim() function.
}
\keyword{export}
