% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFA.R
\name{neg_loglikelihood}
\alias{neg_loglikelihood}
\title{Calculate negative penalized likelihood, used for calls
to the optim() function.}
\usage{
neg_loglikelihood(parameters, modelEnv)
}
\arguments{
\item{parameters}{Vectorized parameter space.}

\item{modelEnv}{Environment variable contains parameter space,
and global variables such as N,G,C,detection matrix B, etc}
}
\value{
Scalar penalized likelihood
}
\description{
The penalized likelihood function:
\eqn{f(A,Z,\beta,O,U) = \sum [lnP(B;A,Z,U,V,\beta,\gamma)]_ij -
\epsilon_1 * ||A||_2^2 - \epsilon_2 * ||Z||_2^2 -
\epsilon_3*||\beta||_2^2 - \epsilon_2 * ||\gamma||_2^2}
}
\keyword{internal}
