% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFA.R
\name{OptimBFA}
\alias{OptimBFA}
\title{Optimize parameters of BFA's likelihood function}
\usage{
OptimBFA(modelEnv, maxit, method)
}
\arguments{
\item{modelEnv}{Environment variable contains parameter space,
and global variables such as N,G,C,detection matrix B, etc}

\item{maxit}{Maximum number of iteration with respect to objective function,
default is 300 iterations}

\item{method}{Optimization method, default is the conjugate gradient approach
L-BFGS-B is recommended for smaller dataset less than 10k cells}
}
\value{
The entire model environment
}
\description{
Optimize parameters of BFA's likelihood function
}
\keyword{internal}
