% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFA.R
\name{restore}
\alias{restore}
\title{Restore the vector of parameter space into their seperated parameterization}
\usage{
restore(parameters, modelEnv)
}
\arguments{
\item{parameters:}{Vectorized parameter space.}

\item{modelEnv:}{Environment variable contains parameter space ,
and global variables such as N,G,C,T,detection matrix B etc}
}
\value{
A list parameters containing the following parameters:
{A,Z,U,V,beta,gamma,epsilon}
}
\description{
Restore the vector of parameter space into their seperated parameterization
}
\keyword{internal}
