% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zinb}
\alias{zinb}
\title{A zinb object after fitting a ZINB-WaVE on cDC/pre-DC dataset}
\format{A ZinbModel object contains all the parameters of ZINB-WaVE}
\source{
\url{https://github.com/drisso/zinbwave}
}
\usage{
zinb
}
\description{
A zinb model object contains all the parameters of ZINB-WaVE model after
fitting dentritic dataset on
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE89232
}
\keyword{datasets}
