% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zinb_toy}
\alias{zinb_toy}
\title{example zinb object after fitting a toy dataset with 5 cells and 10 genes}
\format{A ZinbModel class contains all of its parameters after fitting
a toy dataset}
\usage{
zinb_toy
}
\description{
The toy dataset is generated from the following code
require(zinbwave)
GeneExpr = matrix(rpois(50,1),nrow = 10,ncol = 5)
rownames(GeneExpr) = paste0("gene",seq_len(nrow(GeneExpr)))
colnames(GeneExpr) = paste0("cell",seq_len(ncol(GeneExpr)))
celltype = as.factor(sample(c(1,2,3),5,replace = TRUE))
zinb = zinbFit(Y = GeneExpr,K=2)
}
\keyword{datasets}
