% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celltype}
\alias{celltype}
\title{Cell types as labels of example scRNA-seq dataset(exprdata)}
\format{A vector contains 15 cell types with its length 1059}
\source{
\url{http://imlspenticton.uzh.ch/robinson_lab/conquer/data-mae/GSE48968-GPL13112.rds}
}
\usage{
celltype
}
\description{
A vector contains the cell types  as labels for cells in
example scRNA-seq dataset(exprdata)
}
\keyword{datasets}
