% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprdata}
\alias{exprdata}
\title{example scRNA-seq dataset(exprdata)}
\format{A 500 x 950 matrix with gene counts value}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE89232}
}
\usage{
exprdata
}
\description{
A matrix contains 950 cells and 500 genes. The source of this dataset is cDC/
pre-DC cells(see supplementary files)
We subset most variant 500 genes  as  example scRNA-seq dataset(exprdata)
}
\keyword{datasets}
